Start transition to class and screen libraries
From now on, this game is going to use the middleclass and ScreenManager
Lua libraries.
The game is sort of a mess right now, but is functional. There are bugs
to be fixed still and the menus need to be refined. Hooks also need to
be updated as some of them were removed or ruined.